pyproject.toml
[project]
name = "twio-cairn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.3",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.44.0",
]
